Fix dotfiles install
The Brewfile/Caskfile included old formulaes that I removed to make the
installer work on a new machine. I also modifiied the zshrc to source
the oh-my-zsh stuff using the $HOME variable so it would work across
machines with different usernames.

The powconfig will allow pow to work with my default ruby version.